requirements.txt
jupyterlab
notebook
numpy
pandas
Pillow
opencv-python
pyyaml
scikit-learn
tqdm
aiofiles
aiohttp
asyncio
fastai
torch
torchvision
numpy
starlette
uvicorn
python-multipart
nbformat
nbclient
matplotlib
